requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyjson==0.3.3
APIResponse==0.2
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitarray==2.7.3
bleach==5.0.1
blinker==1.4
boto3==1.26.112
botocore==1.29.112
Brlapi==0.7.0
cache==1.0.3
cachetools==5.2.0
camera==1.3.0
celerybeat-mongo==0.2.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cinemagoer==2022.2.11
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codegen==1.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cupshelpers==1.0
cv==1.0.0
cycler==0.10.0
cytoolz==0.12.1
datasets==2.3.2
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
deepface==0.0.78
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.7.0
django-cron==0.6.0
django-crontab==0.7.1
django-crum==0.7.9
django-extensions==3.1.5
django-filter==22.1
django-layers-hr==1.11.1
django-request-logging==0.7.5
django-rest-swagger==2.2.0
django-social-auth==0.7.28
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangorestframework-yaml==2.0.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
drf-yasg2==1.19.4
duplicity==0.8.12.0
entrypoints==0.4
ephem==3.7.7.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
evaluate==0.1.2
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.0.12
fire==0.5.0
Flask==2.0.0
Flask-Cors==3.0.9
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gdown==4.6.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gunicorn==19.9.0
h5py==3.7.0
helper==2.5.0
hexbytes==0.3.0
httpie==1.0.3
httplib2==0.14.0
huggingface-hub==0.8.1
idna==2.10
imageio==2.6.1
IMDbPY==2022.7.9
importlib-metadata==4.12.0
importlib-resources==5.8.0
imutils==0.5.3
inflection==0.5.1
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jwt==1.3.1
keras==2.9.0
keras-nightly==2.11.0.dev2022082907
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.1.0
kombu==5.2.4
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
libs==0.0.10
lightgbm==3.3.5
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.8
lxml==4.9.0
macaroonbakery==1.3.1
mahotas==1.4.9
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.1
matplotlib-inline==0.1.3
mediapipe==0.9.1.0
mistune==0.8.4
mnist==0.2.2
mongoengine==0.24.2
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mtcnn==0.1.1
multidict==6.0.2
multiprocess==0.70.13
multitasking==0.0.11
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.4
nltk==3.2.4
notebook==6.4.12
npm==0.1.1
numpy==1.24.3
oauth2==1.9.0.post1
oauthlib==3.2.0
olefile==0.46
openapi==1.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.7.0.68
opencv-python==4.6.0.66
opt-einsum==3.3.0
optional-django==0.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.6.0
params==0.9.0
parsimonious==0.9.0
parso==0.8.3
patterns==0.3
pexpect==4.8.0
pickleshare==0.7.5
picture==0.0.1
pictures==0.7
pika==1.1.0
Pillow==8.1.1
pkgutil_resolve_name==1.3.10
ply==3.11
preprocessing==0.1.13
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.17
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
pymongo==3.12.3
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.5
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pyspark==3.3.0
pytesseract==0.3.9
python-apt==2.0.1+ubuntu0.20.4.1
python-barcode==0.14.0
python-crontab==2.7.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.7
python-memcached==1.58
python-openid==2.2.5
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywhatkit==5.4
pywifi==1.1.12
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
razorpay==1.3.0
redis==4.4.0
regex==2022.6.2
reportlab==3.5.34
repoze.lru==0.7
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resources==0.0.1
responses==0.18.0
retina-face==0.0.13
rlp==3.0.0
Routes==2.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
schedule==1.1.0
scikit-image==0.16.2
scikit-learn==1.1.1
scikit-surprise==1.1.1
scipy==1.8.1
scraper==0.1.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
shell==1.0.1
simplejson==3.17.6
sip==6.7.7
six==1.16.0
sklearn==0.0
slug==2.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
SQLAlchemy==1.4.39
sqlparse==0.4.1
stack-data==0.3.0
staty==1.2.4
Strip==0.0.1.dev14
stripe==3.0.0
stripe3==0.0.0
stuff==0.0.0.dev0
surprise==0.1
systemd-python==234
tb-nightly==2.11.0a20220828
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tf==1.0.0
tf-estimator-nightly==2.11.0.dev2022082908
tf-nightly==2.11.0.dev20220828
threadpoolctl==3.1.0
tinycss2==1.1.1
tmdbv3api==1.6.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.7
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.11
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
web3==6.1.0
webencodings==0.5.1
websockets==11.0.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wikipedia==1.4.0
wrapt==1.14.1
xkit==0.0.0
xxhash==3.0.0
yarl==1.7.2
yfinance==0.1.74
zipp==3.8.0
